# server/app/api/swagger.yml
openapi: "3.0.3"
info:
  description: "This is the Databus API documentation."
  version: "2.1.0"
  title: "Databus API"
components:
  securitySchemes:
    ApiKeyAuth:        
      type: apiKey
      in: header       
      name: X-API-KEY  
paths:
...
  /api/publish:
    post:
      summary: "Publish Metadata"
      description: "
      Publish metadata Json-LD (DataID) on the Databus. 

      
      A [DataId](http://dataid.dbpedia.org/ns/core.html) is an RDF metadata document with an entity of rdf:type [dataid:Dataset](http://dataid.dbpedia.org/ns/core#Dataset) at its heart. The document contains information about the corresponding dataset:
      
        * The containing files and their metadata
        
        * A Databus version
        
        * A Databus artifact
        
        * A Databus group
        
      
      The server will try to auto-complete your input and can optionally calculate the required sha256sum and byteSize properties.
      
            
           The input data must be supplied as Json-LD.
           
            
           The input data will be filtered with the following construct queries:
            
                * [Construct group](./server/app/common/queries/constructs/construct-group.sparql) query
              
                * [Construct version](./server/app/common/queries/constructs/construct-version.sparql) query
                
              
           Each construct query result will then be validated with their corresponding SHACL shape (e.g. group construct query result validated by group SHACL shapes):
            
                * [Group SHACL](./server/app/common/shacl/group-shacl.ttl) shapes
              
                * [Version SHACL](./server/app/common/shacl/dataid-shacl.ttl) shapes
      "
      operationId: "publish"
      tags:
        - Main API
      security:
        - ApiKeyAuth: []
      parameters:
        - name: fetch-file-properties
          in: query
          schema:
            type: boolean
          description: "If set to false, the server will not verify the *sha256sum* and *byteSize* properties of any *databus:Part*. Default value is *true*."
          required: false
          example: "true"
        - name: log-level
          in: query
          schema:
            type: string
          description: "Desired detail of the publish report. Can be set to either 'error', 'info' or 'debug'."
          required: false
          example: "info"
      requestBody:
        description: "Metadata in JSONLD. May contain any number of groups, artifacts and datasets"
        required: true
        content:
          application/ld+json:
            schema:
              type: object
              example: |
                {
                  "@context": "https://downloads.dbpedia.org/databus/context.jsonld",
                  "@graph": [
                    {
                      "@type": "Dataset",
                      "@id": "https://dev.databus.dbpedia.org/<your username>/test_group/test_artifact/2022-02-09#Dataset",
                      "hasVersion": "2022-02-09",
                      "title": "test_group test_artifact",
                      "description": "version of the test_artifact dataset from DBpedia",
                      "license": "https://dalicc.net/licenselibrary/Apache-2.0",
                      "distribution": [
                        {
                          "@type": "Part",
                          "formatExtension": "md",
                          "compression": "none",
                          "downloadURL": "https://raw.githubusercontent.com/dbpedia/databus/68f976e29e2db15472f1b664a6fd5807b88d1370/README.md"
                        }
                      ]
                    }
                  ]
                }
      responses:
        "200":
          description: "Accepted"
          content:
            application/json:
              schema:
                type: object
                example: |
                  {
                    "logLevel": "info",
                    "log": [
                      {
                        "resource": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                        "msg": "Successfully published group <%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%>.",
                        "payload": {
                          "@context": "https://downloads.dbpedia.org/databus/context.jsonld",
                          "@id": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                          "@type": "Group",
                          "abstract": "%GROUP% datasets from DBpedia",
                          "description": "%GROUP% datasets from DBpedia",
                          "title": "%GROUP% Data"
                        },
                        "level": "info"
                      },
                      {
                        "resource": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%/%ARTIFACT%",
                        "msg": "Successfully published artifact <%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%/%ARTIFACT%>.",
                        "payload": {
                          "@context": "https://downloads.dbpedia.org/databus/context.jsonld",
                          "@graph": [
                            {
                              "@id": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%/%ARTIFACT%",
                              "@type": "Artifact",
                              "group": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                              "abstract": "%ARTIFACT% data from DBpedia",
                              "description": "%ARTIFACT% data from DBpedia",
                              "title": "%ARTIFACT%"
                            },
                            {
                              "@id": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                              "@type": "Group"
                            }
                          ]
                        },
                        "level": "info"
                      }
                    ]
                  }
        "403":
          description: "Error: Forbidden. The account failed to authenticate."
          content:
            application/json:
              schema:
                type: object
                example: |
                  {
                    "logLevel": "info",
                    "log": [
                      {
                        "resource": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                        "msg": "Not allowed to access namespace of group identifier <%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%>.",
                        "level": "error"
                      }
                    ]
                  }
  /sparql:
    post:
      summary: "Send SPARQL Query"
      description: "Send a SPARQL request to Databus. SPARQL queries are useful for retrieving the published metadata. This gives users flexibility to make complicated requests combining files from different datasets and of different versions."
      operationId: "sparql-post"
      tags:
        - Main API
      security:
        - ApiKeyAuth: [ ]
      requestBody:
        description: "A SPARQL query."
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: "select distinct * where {?a ?b ?c} LIMIT 2"
      responses:
        "200":
          description: "The result of the SPARQL query."
          content:
            application/sparql-results+xml:
              schema:
                type: object
                xml:
                  name: 'sparql'
                example: |
                  <sparql xmlns="http://www.w3.org/2005/sparql-results#" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.w3.org/2001/sw/DataAccess/rf1/result2.xsd">
                    <head>
                      <variable name="a"/>
                      <variable name="b"/>
                      <variable name="c"/>
                    </head>
                    <results distinct="false" ordered="true">
                      <result>
                        <binding name="a">
                          <uri>http://www.openlinksw.com/schemas/virtrdf#DefaultQuadMap</uri>
                        </binding>
                        <binding name="b">
                          <uri>http://www.w3.org/1999/02/22-rdf-syntax-ns#type</uri>
                        </binding>
                        <binding name="c">
                          <uri>http://www.openlinksw.com/schemas/virtrdf#QuadMap</uri>
                        </binding>
                      </result>
                      <result>
                        <binding name="a">
                          <uri>http://www.openlinksw.com/schemas/virtrdf#DefaultServiceMap</uri>
                        </binding>
                        <binding name="b">
                          <uri>http://www.w3.org/1999/02/22-rdf-syntax-ns#type</uri>
                        </binding>
                        <binding name="c">
                          <uri>http://www.openlinksw.com/schemas/virtrdf#QuadMap</uri>
                        </binding>
                      </result>
                    </results>
                  </sparql>
        "400":
          description: "Error: Bad Request"
          content:
            text/plain:
              schema:
                type: string
                example: |
                  Virtuoso 37000 Error SP030: SPARQL compiler, line 2: syntax error at '=' before 'distinct'

                  SPARQL query:
                  define sql:big-data-const 0
                  select= distinct * where {?a ?b ?c} LIMIT 2
  /api/search:
    get:
      summary: "Search"
      description: "Search the index of the Databus. For a full list of the search syntax please refer to the [Lookup Documentation](https://github.com/dbpedia/dbpedia-lookup)"
      operationId: "search"
      tags:
        - Main API
      parameters:
        - name: query
          in: query
          description: search query
          schema:
            type: string
          required: true
          example: '%ARTIFACT%'
        - name: typeName
          in: query
          description: The type of entities to return. White space separated list of one or multiples of
            * Account
            * Group
            * Artifact
            * Version
            * Collection
          schema:
            type: string
          required: false
          example: 'Artifact'
        - name: part
          in: query
          description: A query for one or multiple parta of the entity uri (white space separated list).
          schema:
            type: string
          required: false
          example: ''
        - name: partRequired
          in: query
          description: If set to true, results without a match on the part field are omitted. 
          schema:
            type: boolean
          required: false
          example: 'true'
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                example: '{
                  "docs": [
                    {
                      "typeName": [
                        "Artifact"
                      ],
                      "score": [
                        "337.63354"
                      ],
                      "label": [
                        "%GROUP%s %ARTIFACT%"
                      ],
                      "resource": [
                        "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%/%ARTIFACT%"
                      ],
                      "part": [
                        "%USERNAME%",
                        "%GROUP%",
                        "%ARTIFACT%"
                      ]
                    }
                  ],
                  "query": "%ARTIFACT%"
                  
                }'
        "404":
          description: "Error: Not Found."
...